sources and "extra" support. Keep a track of source packages, only process their build-depends once (an efficiency improvement) and write sources lists for each set's seed & dependencies and build-dependencies.
Split the simpler "is build_depend" logic into three discreet settings
each of which implies the ones below it in the list:

    build_depend
	This package is a physical build dependency of another
	(specified by pkg).

    build_tree
	This package and its dependencies should be added to the
	build_depends list rather than the depends list.

    second_class
	This package and its dependencies will not be removed from
	the build_depends list (and by effect promoted into the depends
	list) if they already exist there instead they will be skipped.

Splitting the first two corrects the bug that a dependency of a
build-dependency would still be listed as a build-dependency in the Why
field.

Adding the third creates support for seeds that don't steal from the
build-depends lists of earlier ones to populate their own depends
list.


Across all the sets, for each source package add any generated binary
packages not already placed in a set to a second-class "extra" set.
Don't write out a sources list for this set (just build-sources) as
it is by inference always empty.


Only write the seed and depends to the "supported" file (matching
"base" and "desktop"), this used to be the "supported.only" file.

Write the combination of the primary sets (base, desktop, supported)
including build-depends to the new "all" file, which has a companion
"all.sources"

Write the complete lists (including "extra") to "all+extra" and
"all+extra.sources"