Initial support for specifying snaps in seeds
This support more or less passes through entries of the form

  * snap:foo [arch list]

or

  * snap:foo/classic [arch list]

(for classic snaps)

into a file "seedname.snaps". No checks on availability are performed -
it is the responsibility of the seed's maintainers to ensure that the
snap is present for all seeded arches.

Output files which previously existed are unaffected, so consumers do
not have to worry about lists they are using starting to contain non-deb
packages, as long as they do not end up parsing the new .snaps files.